# [XEB] Faster and more extensible simulation