This PR speeds up the classical simulation of XEB circuits, which is important because it is in the hot loop for optimization. 

Single threaded: old: 0.72s new: 3.72s speedup: 5x
multiprocessing pool on my laptop: old: 1.42s new: 0.21s speedup: 7x

- Instead of using a loosely schema'd dictionary for passing arguments to the simulation helper function, use a private dataclass
- The helper function has been promoted to a helper closure class. This is so 1) it can work under multiprocessing while 2) storing its `self.Simulator` to use. This enables a user-provided simulator object. Anything that follows the `SimulatesIntermediateState` interface. Currently, this is just `cirq.Simulator` but there's an outstanding request to use qsim in this way. https://github.com/quantumlib/qsim/issues/101. Please see the linked issue for potential speedups. Right now, `cirq.Simulator` with intermediate capabilities is faster than qsim operated naively